# optimize Array.of and Array.from for cpu pipelining and branch prediction