passes 262 webkit tests (`Tools/Scripts/test262-runner --release`, build first)
todo: check cpu cycles and instructions count (needs bun build for @mitata/counters)

patch: https://gist.github.com/evanwashere/17778413aba8b1eec1adfe5c6ef04a69

before
```js
------------------------------------------------------- -------------------------------
Array.of(512)                              1.58 µs/iter   1.61 µs   1.63 µs ▁▃▆▇▅▂▂▄█▆▂
Array.of(4096)                            10.23 µs/iter  10.23 µs  10.28 µs ▅▃▃▅██▅▁▁▁▃
Array.of(32768)                           87.47 µs/iter  92.46 µs 110.21 µs ▁█▄▂▁▁▁▃▁▁▁
Array.of(65536)                          154.52 µs/iter 156.46 µs 170.75 µs ▁█▇▇▄▃▂▁▁▁▁

------------------------------------------------------- -------------------------------
Array.from(arr 512)                      196.79 ns/iter 199.10 ns 224.52 ns ▂▅█▇▄▂▁▁▁▁▁
Array.from(arr 4096)                     732.59 ns/iter 734.55 ns 835.02 ns ▁▅█▂▁▁▁▁▁▁▁
Array.from(arr 32768)                      5.88 µs/iter   5.90 µs   5.93 µs ▂▂▂▅▄▄█▄▅▄▂
Array.from(arr 65536)                     11.11 µs/iter  11.65 µs  11.69 µs █▂▁▁▁▁▁▁▁▃█

------------------------------------------------------- -------------------------------
Array.fromAsync(arr 512)                 131.26 ns/iter 131.90 ns 245.42 ns ▂█▄▂▁▁▁▁▁▁▁
Array.fromAsync(arr 4096)                 10.04 µs/iter  11.88 µs  13.33 µs ▄▃█▄▃▁▁▂▄▃▂
Array.fromAsync(arr 32768)                 9.71 µs/iter  10.25 µs  13.33 µs ▅██▆▅▂▁▃▂▁▃
Array.fromAsync(arr 65536)                 9.97 µs/iter  11.83 µs  13.00 µs ▇██▅▂▄▁▂▄▇▂

------------------------------------------------------- -------------------------------
Array.from(512)                           47.93 µs/iter  74.00 µs  76.25 µs █▃▁▁▁▁▁▂▃▆█
Array.from(4096)                          54.00 µs/iter  82.88 µs  87.17 µs ▆▃▂▂▁▁▁▁▃█▃
Array.from(32768)                        105.18 µs/iter 160.21 µs 181.71 µs █▃▁▁▁▁▂▂▂▃▃
Array.from(65536)                        184.21 µs/iter 218.63 µs 239.04 µs ▆▃▁▃▁▂▁▂█▄▃

------------------------------------------------------- -------------------------------
Array.from(512, () => 1)                   8.46 µs/iter  10.50 µs  12.62 µs ▆█▅▃▁▁▁▅▂▁▂
Array.from(4096, () => 1)                 25.36 µs/iter  27.08 µs  29.75 µs ██▃▃▆▆█▆▃▃▃
Array.from(32768, () => 1)               125.75 µs/iter 128.83 µs 133.63 µs ▄▂▄█▂▄▄▄▁▅▂
Array.from(65536, () => 1)               238.31 µs/iter 246.00 µs 253.54 µs ▅█▅▅▅▁▃▆▃▅▃

------------------------------------------------------- -------------------------------
Array.from(iterator 512)                  67.29 µs/iter  97.17 µs 112.63 µs █▅▂▁▁▁▁▄▅▅▃
Array.from(iterator 4096)                141.72 µs/iter 165.42 µs 184.96 µs ▅▅▃▁▁▁▂▄█▄▂
Array.from(iterator 32768)               250.42 µs/iter 248.21 µs 438.42 µs ▃▃▇█▃▂▁▂▁▁▃
Array.from(iterator 65536)               375.99 µs/iter 378.46 µs 730.96 µs ▄▅█▅▁▁▁▁▁▁▂

------------------------------------------------------- -------------------------------
Array.fromAsync(iterator 512)            485.39 µs/iter  87.29 µs   4.85 ms █▁▁▁▁▁▁▁▁▁▁
Array.fromAsync(iterator 4096)           573.32 µs/iter 579.96 µs 610.75 µs ▂▃▃▄█▅▃▂▂▁▁
Array.fromAsync(iterator 32768)            4.28 ms/iter   4.29 ms   4.42 ms ▂▅▇█▃▂▂▂▁▁▁
Array.fromAsync(iterator 65536)            8.50 ms/iter   8.52 ms   8.70 ms ▁▁▃█▇▃▁▁▁▁▂

------------------------------------------------------- -------------------------------
Array.from(iterator 512, () => 1)          2.62 µs/iter   2.60 µs   3.11 µs █▂▁▁▁▁▁▁▁▁▁
Array.from(iterator 4096, () => 1)        18.68 µs/iter  18.61 µs  18.88 µs ▂█▄▄▁▁▁▁▁▁▂
Array.from(iterator 32768, () => 1)      168.15 µs/iter 169.92 µs 182.17 µs ▁▁▁▂▃▂█▃▂▁▁
Array.from(iterator 65536, () => 1)      306.85 µs/iter 311.50 µs 324.46 µs ▁▁▁▂▃▆█▅▃▂▁

------------------------------------------------------- -------------------------------
Array.fromAsync(iterator 512, () => 1)   108.26 µs/iter 110.75 µs 131.71 µs ▁▂█▃▂▂▂▁▁▁▁
Array.fromAsync(iterator 4096, () => 1)  753.81 µs/iter 760.75 µs 788.00 µs ▂▂▂▃█▇▅▃▂▂▁
Array.fromAsync(iterator 32768, () => 1)   5.69 ms/iter   5.70 ms   5.79 ms ▂▅███▄▂▂▂▂▁
Array.fromAsync(iterator 65536, () => 1)  11.34 ms/iter  11.36 ms  11.70 ms ▃█▆▃▁▂▁▁▁▁▁
```

after
```js
------------------------------------------------------- -------------------------------
Array.of(512)                              1.45 µs/iter   1.49 µs   1.50 µs ▁▁▅▇▂▂▂▁▂█▃
Array.of(4096)                             9.17 µs/iter   9.18 µs   9.23 µs ▂▁▁▄█▅▇▁▅▁▂
Array.of(32768)                           73.83 µs/iter  74.01 µs  74.57 µs █▃▃▁▆▁▆▃▁▁▃
Array.of(65536)                          135.46 µs/iter 135.25 µs 151.42 µs ▁▁▂█▂▂▁▁▁▁▁

------------------------------------------------------- -------------------------------
Array.from(arr 512)                      189.49 ns/iter 190.93 ns 218.66 ns ▁▇█▄▂▁▁▁▁▁▁
Array.from(arr 4096)                     731.71 ns/iter 733.40 ns 852.39 ns ▂█▄▁▁▁▁▁▁▁▁
Array.from(arr 32768)                      5.80 µs/iter   5.80 µs   5.87 µs ▂▅█▅▃▂▁▂▁▁▂
Array.from(arr 65536)                     11.62 µs/iter  11.63 µs  11.64 µs ▂▁▂▂▂▁▄█▄▂▂

------------------------------------------------------- -------------------------------
Array.fromAsync(arr 512)                 130.52 ns/iter 129.05 ns 244.67 ns ▃█▄▁▁▁▁▁▁▁▁
Array.fromAsync(arr 4096)                 10.44 µs/iter  11.29 µs  13.46 µs ▅▄▄█▇▇▇▄▂▄▂
Array.fromAsync(arr 32768)                 9.06 µs/iter  10.08 µs  11.50 µs ▂█▂█▅▂▃▂▃▃▃
Array.fromAsync(arr 65536)                 9.32 µs/iter   9.79 µs  12.75 µs ▅▇█▃▃▃▃▁▁▁▃

------------------------------------------------------- -------------------------------
Array.from(512)                           25.83 µs/iter  53.87 µs  70.96 µs █▂▁▁▁▁▁▁▁▃▂
Array.from(4096)                          61.17 µs/iter  78.58 µs  84.42 µs ▅▂▂▁▁▁▂▃█▆▄
Array.from(32768)                         74.08 µs/iter 124.71 µs 138.42 µs █▆▂▁▁▁▁▁▂▂▅
Array.from(65536)                        107.06 µs/iter 153.58 µs 180.58 µs █▄▂▂▂▁▁▂▃▁▂

------------------------------------------------------- -------------------------------
Array.from(512, () => 1)                   7.76 µs/iter   8.08 µs  10.08 µs ▃▃▂█▃▅▁▂▃▁▂
Array.from(4096, () => 1)                 25.13 µs/iter  26.92 µs  31.58 µs █▂▂▂▄▇▄▂▄▁▂
Array.from(32768, () => 1)               116.97 µs/iter 120.29 µs 126.92 µs ▆██▅▃▃▃▅▃▁▃
Array.from(65536, () => 1)               219.08 µs/iter 222.67 µs 231.50 µs ▂█▁█▁▇▄▄▁▂▂

------------------------------------------------------- -------------------------------
Array.from(iterator 512)                  64.47 µs/iter  88.71 µs 102.08 µs ▅▅▂▂▃▁▁▃█▄▂
Array.from(iterator 4096)                102.48 µs/iter 123.96 µs 146.96 µs █▂▂▃▁▁▃▇▄▁▄
Array.from(iterator 32768)               226.90 µs/iter 232.63 µs 398.33 µs ▅▇▃██▁▁▁▁▁▅
Array.from(iterator 65536)               297.57 µs/iter 322.12 µs 350.96 µs ▇▅▂▁▁▂█▇▃▁▃

------------------------------------------------------- -------------------------------
Array.fromAsync(iterator 512)            430.12 µs/iter  84.46 µs   4.66 ms █▁▁▁▁▁▁▁▁▁▁
Array.fromAsync(iterator 4096)           559.06 µs/iter 563.71 µs 588.04 µs ▁▁▃██▅▃▃▂▁▁
Array.fromAsync(iterator 32768)            4.10 ms/iter   4.11 ms   4.18 ms ▁▂▆█▅▂▁▁▁▁▁
Array.fromAsync(iterator 65536)            8.15 ms/iter   8.19 ms   8.31 ms ▃█▅▂▂▃▃▁▂▁▁

------------------------------------------------------- -------------------------------
Array.from(iterator 512, () => 1)          2.17 µs/iter   2.13 µs   2.87 µs █▁▁▁▁▁▁▁▁▁▁
Array.from(iterator 4096, () => 1)        15.02 µs/iter  15.00 µs  15.12 µs ██▆▁▃▁▁▁▁▁▃
Array.from(iterator 32768, () => 1)      134.04 µs/iter 136.75 µs 150.92 µs ▁▁▂█▃▆▃▂▂▁▁
Array.from(iterator 65536, () => 1)      241.85 µs/iter 244.96 µs 258.21 µs ▁▁▂▂█▄▅▃▂▁▁

------------------------------------------------------- -------------------------------
Array.fromAsync(iterator 512, () => 1)   107.68 µs/iter 109.46 µs 130.00 µs ▁▂█▃▂▂▁▁▁▁▁
Array.fromAsync(iterator 4096, () => 1)  754.21 µs/iter 760.25 µs 791.71 µs ▁▁▁▃█▅▄▃▂▁▁
Array.fromAsync(iterator 32768, () => 1)   5.64 ms/iter   5.64 ms   5.70 ms ▁▆██▄▃▃▂▁▁▁
Array.fromAsync(iterator 65536, () => 1)  11.22 ms/iter  11.23 ms  11.39 ms ▂▆█▄▃▁▁▁▁▁▁
```